Reformat certs
This reformats the textual version of the cert files to match what the
current output of the tools on my machine is.  Nothing that reads the
cert files reads the text format, they only read the encoded version
at the top of the file, so this should be a no-op.

Test: manual inspection
Change-Id: Idd20ab5f71e698f79d93cf5091d63a2474508364